requirements.txt
altgraph==0.17.2
appdirs==1.4.4
asgiref==3.4.1
awsebcli==3.20.3
boto3==1.20.27
botocore==1.23.52
cement==2.8.2
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.11
colorama==0.4.3
coreapi==2.3.3
coreschema==0.0.4
crispy-forms-materialize==0.2
cryptography==36.0.1
defusedxml==0.7.1
diff-match-patch==20200713
distlib==0.3.1
Django==3.2.7
django-bootstrap-datepicker-plus @ file:///C:/Users/chris/Desktop/django_bootstrap_datepicker_plus-4.0.0-py3-none-any.whl
django-cors-headers==3.11.0
django-crispy-forms==1.13.0
django-filter==21.1
django-import-export==2.7.1
django-phonenumber-field==6.0.0
django-phonenumbers==1.0.1
django-rest-knox==4.2.0
django-ses==2.3.1
django-templated-mail==1.1.1
djangorestframework==3.13.1
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
et-xmlfile==1.1.0
fbs==1.0.1
filelock==3.0.12
future==0.16.0
idna==3.3
importlib-metadata==4.10.0
itypes==1.2.0
Jinja2==3.0.3
jmespath==0.10.0
macholib==1.15.2
Markdown==3.3.6
MarkupPy==1.14
MarkupSafe==2.0.1
oauthlib==3.2.0
odfpy==1.4.1
openpyxl==3.0.9
pathspec==0.9.0
pefile==2021.9.3
phonenumbers==8.12.41
Pillow==9.0.0
pycparser==2.21
Pygments==2.11.1
PyInstaller==3.4
PyJWT==2.3.0
pypiwin32==223
PyQt5==5.15.6
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.0
PyQt6==6.2.1
PyQt6-Qt6==6.2.1
PyQt6-sip==13.1.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2021.1
pywin32==303
pywin32-ctypes==0.2.0
PyYAML==5.4.1
requests==2.26.0
requests-oauthlib==1.3.1
s3transfer==0.5.0
semantic-version==2.8.5
six==1.14.0
social-auth-app-django==4.0.0
social-auth-core==4.2.0
sqlparse==0.4.1
stripe==2.65.0
tablib==3.2.0
termcolor==1.1.0
uritemplate==4.1.1
urllib3==1.26.7
virtualenv==20.4.7
wcwidth==0.1.9
xlrd==2.0.1
xlwt==1.3.0
zipp==3.7.0
